internal/xhttp: move memstat handler from plugin svr to shared xhttp package and make GC opt-in
