+++ cmd/yokecd/internal/svr/run.go
@@ -10,14 +10,11 @@ import (
 	"log/slog"
 	"net/http"
 	"os"
-	"runtime"
 	"strings"
 	"sync"
 	"time"
 
 	"github.com/davidmdm/conf"
-
-	units "github.com/docker/go-units"
 
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
@@ -135,39 +132,10 @@ type ExecResponse struct {
 	Stderr string
 }
 
-type HumanMemStats struct {
-	TotalAlloc string
-	Sys        string
-	HeapAlloc  string
-	HeapSys    string
-	HeapIdle   string
-	HeapInuse  string
-	NextGC     string
-}
-
-func humanSize(value uint64) string {
-	return units.HumanSize(float64(value))
-}
-
 func Handler(mods *cache.ModuleCache, logger *slog.Logger, client *k8s.Client) http.Handler {
 	mux := http.NewServeMux()
 
-	mux.HandleFunc("GET /memstats", func(w http.ResponseWriter, r *http.Request) {
-		runtime.GC()
-
-		var stats runtime.MemStats
-		runtime.ReadMemStats(&stats)
-
-		_ = json.NewEncoder(w).Encode(HumanMemStats{
-			TotalAlloc: humanSize(stats.TotalAlloc),
-			Sys:        humanSize(stats.Sys),
-			HeapAlloc:  humanSize(stats.HeapAlloc),
-			HeapSys:    humanSize(stats.HeapSys),
-			HeapIdle:   humanSize(stats.HeapIdle),
-			HeapInuse:  humanSize(stats.HeapInuse),
-			NextGC:     humanSize(stats.NextGC),
-		})
-	})
+	mux.HandleFunc("GET /memstats", xhttp.MemStatHandler)
 
 	mux.HandleFunc("POST /exec", func(w http.ResponseWriter, r *http.Request) {
 		var ex ExecuteReq
